+++ test/helpers.dart
@@ -4,6 +4,8 @@ import 'package:secretary/secretary.dart';
 final retryPredicate = predicate<SecretaryEvent>((e) => e.isRetry);
 final failurePredicate = predicate<SecretaryEvent>((e) => e.isFailure);
 final successPredicate = predicate<SecretaryEvent>((e) => e.isSuccess);
+final recurringFinishedPredicate =
+    predicate<SecretaryEvent>((e) => e is RecurringTaskFinishedEvent);
 
 Matcher hasKey<K, T>(K key) =>
     predicate<SecretaryEvent<K, T>>((e) => e.key == key);